requirements.txt
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
Flask==3.0.3
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
graphviz==0.20.3
httplib2==0.22.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
mutagen==1.47.0
numpy==2.1.2
oauthlib==3.2.2
opencv-python==4.10.0.84
Pillow==9.0.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
pyparsing==3.2.0
pytube==15.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.2.0
tk==0.1.0
tqdm==4.66.5
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.3
yt-dlp==2024.10.7
